Migrate eval cfgs commands to config module
